Rework the Batch API interface to be thread-safe.
